+++ PulsarEngine/Network/PacketExpansion.hpp
@@ -124,7 +124,7 @@ struct PulSELECT : public RKNet::SELECTPacket {
     u8 blockedTrackCount;  // Number of valid entries in blockedTracks
     u8 curBlockingArrayIdx;  // Current write index in circular buffer
     bool lastGroupedTrackPlayed;  // Whether most recent track was a grouped track
-    u8 characterTables;  // two bits per local hud slot, up to two local slots
+    u16 characterTables;  // six bits per local hud slot, up to two local slots
     u16 blockedTracks[12];  // PulsarId array (up to MAX_TRACK_BLOCKING tracks)
 
     // Anti-cheat verification tag - proves sender has correct encryption key
